+++ Sources/TipoVetor.c
@@ -4,7 +4,7 @@ void AlocaVetor(TipoVetor* vetor, unsigned int TamanhoVetor){
   vetor->ListaIdMatriz = (TipoMatriz*) malloc (TamanhoVetor*sizeof(TipoMatriz));
 }
 
-void BubbleSort (TipoVetor *Vetor, int TamanhoVetor, int *Compara, int *Movimenta){
+void BubbleSort (TipoVetor *Vetor, int TamanhoVetor, long int *Compara, long int *Movimenta){
   int i,j;
   TipoMatriz aux;
 
@@ -22,7 +22,7 @@ void BubbleSort (TipoVetor *Vetor, int TamanhoVetor, int *Compara, int *Moviment
   }
 }
 
-void SelectSort (TipoVetor *Vetor, int TamanhoVetor, int *Compara, int *Movimenta){
+void SelectSort (TipoVetor *Vetor, int TamanhoVetor, long int *Compara, long int *Movimenta){
   int i,j,min;
   TipoMatriz aux;
 
@@ -41,7 +41,7 @@ void SelectSort (TipoVetor *Vetor, int TamanhoVetor, int *Compara, int *Moviment
   }
 }
 
-void InsertSort (TipoVetor *Vetor, int TamanhoVetor, int *Compara, int *Movimenta){
+void InsertSort (TipoVetor *Vetor, int TamanhoVetor, long int *Compara, long int *Movimenta){
   int i,j;
   TipoMatriz aux;
 
@@ -59,7 +59,7 @@ void InsertSort (TipoVetor *Vetor, int TamanhoVetor, int *Compara, int *Moviment
   }
 }
 
-void ShellSort (TipoVetor *Vetor, int TamanhoVetor, int *Compara, int *Movimenta){
+void ShellSort (TipoVetor *Vetor, int TamanhoVetor, long int *Compara, long int *Movimenta){
   int i,j,h=1;
   TipoMatriz aux;
 
@@ -83,7 +83,7 @@ void ShellSort (TipoVetor *Vetor, int TamanhoVetor, int *Compara, int *Movimenta
   } while(h != 1);
 }
 
-void Particao (int esq,int dir,int* i, int* j, TipoVetor *a, int *Compara, int *Movimenta){
+void Particao (int esq,int dir,int* i, int* j, TipoVetor *a, long int *Compara, long int *Movimenta){
   TipoMatriz pivo,aux;
   *i=esq;
   *j=dir;
@@ -109,7 +109,7 @@ void Particao (int esq,int dir,int* i, int* j, TipoVetor *a, int *Compara, int *
   }while(*i<*j);
 }
 
-void Ordena (int e, int d, TipoVetor *A, int *Compara, int *Movimenta){
+void Ordena (int e, int d, TipoVetor *A, long int *Compara, long int *Movimenta){
   int i,j;
   Particao(e,d,&i,&j,A, Compara, Movimenta);
   if(e<j) Ordena(e,j,A, Compara, Movimenta);
@@ -117,12 +117,12 @@ void Ordena (int e, int d, TipoVetor *A, int *Compara, int *Movimenta){
 
 }
 
-void QuickSort (TipoVetor *Vetor, int TamanhoVetor, int *Compara, int *Movimenta){
+void QuickSort (TipoVetor *Vetor, int TamanhoVetor, long int *Compara, long int *Movimenta){
   *Compara = *Compara+1;
   Ordena(0, TamanhoVetor-1, Vetor, Compara, Movimenta);
 }
 
-void HeapSort (TipoVetor *Vetor, int TamanhoVetor, int *Compara, int *Movimenta){
+void HeapSort (TipoVetor *Vetor, int TamanhoVetor, long int *Compara, long int *Movimenta){
   int i = TamanhoVetor/2;
   int pai, filho;
